--- Engine/Source/logging/src/entry.rs
@@ -1,55 +1,0 @@
-use std::fmt;
-use time::OffsetDateTime;
-
-/// Mirrors `tracing::Level` with an owned, filterable representation.
-/// The discriminant ordering matches severity: `Error = 0` is most severe.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub enum LogLevel {
-    Error = 0,
-    Warn = 1,
-    Info = 2,
-    Debug = 3,
-    Trace = 4,
-}
-
-impl From<&tracing::Level> for LogLevel {
-    fn from(level: &tracing::Level) -> Self {
-        match *level {
-            tracing::Level::ERROR => Self::Error,
-            tracing::Level::WARN => Self::Warn,
-            tracing::Level::INFO => Self::Info,
-            tracing::Level::DEBUG => Self::Debug,
-            tracing::Level::TRACE => Self::Trace,
-        }
-    }
-}
-
-impl fmt::Display for LogLevel {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(
-            f,
-            "{}",
-            match self {
-                Self::Error => "ERROR",
-                Self::Warn => "WARN",
-                Self::Info => "INFO",
-                Self::Debug => "DEBUG",
-                Self::Trace => "TRACE",
-            }
-        )
-    }
-}
-
-/// A single captured log event with full metadata.
-#[derive(Debug, Clone)]
-pub struct LogEntry {
-    /// Severity level.
-    pub level: LogLevel,
-    /// Category (from the `target:` directive or a `category` field; falls back
-    /// to the tracing target, which defaults to the Rust module path).
-    pub category: String,
-    /// UTC timestamp at the moment the event was recorded.
-    pub timestamp: OffsetDateTime,
-    /// Formatted log message.
-    pub message: String,
-}
